refactor(libsinsp): split threadinfo and thread manager components
