+++ userspace/libsinsp/metrics_collector.h
@@ -20,7 +20,7 @@ limitations under the License.
 
 #include <libscap/metrics_v2.h>
 #include <libscap/scap_machine_info.h>
-#include <libsinsp/threadinfo.h>
+#include <libsinsp/thread_manager.h>
 #include <libscap/strl.h>
 #include <cmath>
 #include <memory>
