+++ userspace/libsinsp/thread_group_info.h
@@ -16,6 +16,8 @@ limitations under the License.
 
 */
 
+#pragma once
+
 #include <memory>
 #include <stdint.h>
 #include <list>
